requirements.txt
docker==3.7.2
pyyaml==5.1
jinja2==2.10.1
